chore: rename component files to kebab-case
