+++ apps/web/components/sections/project/overview/key-highlights.tsx
