+++ apps/web/components/sections/project/overview/lead-investors.tsx
